requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
flask-compress>=1.15
langchain>=0.1.20
langchain-core>=0.1.52
langchain-community>=0.0.38
langchain-groq>=0.1.3
langchain-chroma>=0.1.1
langchain-text-splitters>=0.0.1
langchain-huggingface>=0.0.3
langgraph>=0.0.62
chromadb>=0.4.24
sentence-transformers>=2.7.0
rank-bm25>=0.2.2
pypdf>=4.2.0
groq>=0.5.0
sqlalchemy>=2.0.29
python-dotenv>=1.0.1
tiktoken>=0.7.0
pysqlite3-binary>=0.5.2; sys_platform == 'linux'
langchain-community>=0.0.38
ddgs>=0.1.0
